move cal_gdmx and cal_gdmx_k out of LCAO_deepks_pdm
