+++ source/module_hamilt_lcao/module_deepks/cal_gdmx_k.cpp
@@ -1,0 +1,248 @@
+#ifdef __DEEPKS
+
+#include "LCAO_deepks.h"
+#include "module_base/vector3.h"
+#include "module_base/timer.h"
+#include "module_base/constants.h"
+#include "module_hamilt_lcao/module_hcontainer/atom_pair.h"
+#include "module_base/libm/libm.h"
+
+void LCAO_Deepks::cal_gdmx_k(const std::vector<std::vector<std::complex<double>>>& dm,
+    const UnitCell &ucell,
+    const LCAO_Orbitals &orb,
+    Grid_Driver& GridD,
+    const int nks,
+    const std::vector<ModuleBase::Vector3<double>> &kvec_d,
+    const bool isstress)
+{
+    ModuleBase::TITLE("LCAO_Deepks", "cal_gdmx_k");
+    ModuleBase::timer::tick("LCAO_Deepks","cal_gdmx_k");
+
+    const int size = (2 * lmaxd + 1) * (2 * lmaxd + 1);
+
+    for (int iat = 0;iat < ucell.nat;iat++)
+    {
+        for (int inl = 0;inl < inlmax;inl++)
+        {
+            ModuleBase::GlobalFunc::ZEROS(gdmx[iat][inl], size);
+            ModuleBase::GlobalFunc::ZEROS(gdmy[iat][inl], size);
+            ModuleBase::GlobalFunc::ZEROS(gdmz[iat][inl], size);
+        }
+    }
+
+    if (isstress)
+    {
+        for (int ipol = 0;ipol < 6;ipol++)
+        {
+            for (int inl = 0;inl < inlmax;inl++)
+            {
+                ModuleBase::GlobalFunc::ZEROS(gdm_epsl[ipol][inl], size);
+            }
+        }
+    }
+
+    const double Rcut_Alpha = orb.Alpha[0].getRcut();
+    int nrow = this->pv->nrow;
+    for (int T0 = 0; T0 < ucell.ntype; T0++)
+    {
+		Atom* atom0 = &ucell.atoms[T0]; 
+        for (int I0 =0; I0< atom0->na; I0++)
+        {
+            const int iat = ucell.itia2iat(T0,I0);//on which alpha is located
+            const ModuleBase::Vector3<double> tau0 = atom0->tau[I0];
+            GridD.Find_atom(ucell, atom0->tau[I0] ,T0, I0);
+
+            for (int ad1=0; ad1<GridD.getAdjacentNum()+1 ; ++ad1)
+            {
+                const int T1 = GridD.getType(ad1);
+                const int I1 = GridD.getNatom(ad1);
+                const int ibt1 = ucell.itia2iat(T1,I1); //on which chi_mu is located
+                const int start1 = ucell.itiaiw2iwt(T1, I1, 0);
+                
+                const ModuleBase::Vector3<double> tau1 = GridD.getAdjacentTau(ad1);
+				const Atom* atom1 = &ucell.atoms[T1];
+				const int nw1_tot = atom1->nw*GlobalV::NPOL;
+				const double Rcut_AO1 = orb.Phi[T1].getRcut();
+
+                ModuleBase::Vector3<double> dR1(GridD.getBox(ad1).x, GridD.getBox(ad1).y, GridD.getBox(ad1).z); 
+
+				for (int ad2=0; ad2 < GridD.getAdjacentNum()+1 ; ad2++)
+				{
+					const int T2 = GridD.getType(ad2);
+					const int I2 = GridD.getNatom(ad2);
+					const int start2 = ucell.itiaiw2iwt(T2, I2, 0);
+                    const int ibt2 = ucell.itia2iat(T2,I2);
+					const ModuleBase::Vector3<double> tau2 = GridD.getAdjacentTau(ad2);
+					const Atom* atom2 = &ucell.atoms[T2];
+					const int nw2_tot = atom2->nw*GlobalV::NPOL;
+                    ModuleBase::Vector3<double> dR2(GridD.getBox(ad2).x, GridD.getBox(ad2).y, GridD.getBox(ad2).z);
+					
+					const double Rcut_AO2 = orb.Phi[T2].getRcut();
+                	const double dist1 = (tau1-tau0).norm() * ucell.lat0;
+                	const double dist2 = (tau2-tau0).norm() * ucell.lat0;
+
+					if (dist1 > Rcut_Alpha + Rcut_AO1
+							|| dist2 > Rcut_Alpha + Rcut_AO2)
+					{
+						continue;
+					}
+
+                    double r0[3];
+                    double r1[3];
+                    if(isstress)
+                    {
+                        r1[0] = ( tau1.x - tau0.x) ;
+                        r1[1] = ( tau1.y - tau0.y) ;
+                        r1[2] = ( tau1.z - tau0.z) ;
+                        r0[0] = ( tau2.x - tau0.x) ;
+                        r0[1] = ( tau2.y - tau0.y) ;
+                        r0[2] = ( tau2.z - tau0.z) ;
+                    }
+
+                    auto row_indexes = pv->get_indexes_row(ibt1);
+                    auto col_indexes = pv->get_indexes_col(ibt2);
+                    if(row_indexes.size() * col_indexes.size() == 0) continue;
+
+                    hamilt::AtomPair<double> dm_pair(ibt1, ibt2, (dR2-dR1).x, (dR2-dR1).y, (dR2-dR1).z, pv);
+                    dm_pair.allocate(nullptr, 1);
+                    for(int ik=0;ik<nks;ik++)
+                    {
+                        const double arg = - (kvec_d[ik] * (dR2-dR1) ) * ModuleBase::TWO_PI;
+                        double sinp, cosp;
+                        ModuleBase::libm::sincos(arg, &sinp, &cosp);
+                        const std::complex<double> kphase = std::complex<double>(cosp, sinp);
+                        if(ModuleBase::GlobalFunc::IS_COLUMN_MAJOR_KS_SOLVER())
+                        {
+                            dm_pair.add_from_matrix(dm[ik].data(), pv->get_row_size(), kphase, 1);
+                        }
+                        else
+                        {
+                            dm_pair.add_from_matrix(dm[ik].data(), pv->get_col_size(), kphase, 0);
+                        }
+                    }
+                    const double* dm_current = dm_pair.get_pointer();
+
+                    key_tuple key_1(ibt1,dR1.x,dR1.y,dR1.z);
+                    key_tuple key_2(ibt2,dR2.x,dR2.y,dR2.z);
+					for (int iw1l = 0; iw1l < row_indexes.size(); ++iw1l)
+                    {
+                        for (int iw2l = 0; iw2l < col_indexes.size(); ++iw2l)
+                        {
+                            std::vector<double> nlm1 = this->nlm_save_k[iat][key_1][row_indexes[iw1l]][0];
+                            std::vector<std::vector<double>> nlm2 = this->nlm_save_k[iat][key_2][col_indexes[iw2l]];
+
+                            assert(nlm1.size()==nlm2[0].size());
+
+                            int ib=0;
+                            for (int L0 = 0; L0 <= orb.Alpha[0].getLmax();++L0)
+                            {
+                                for (int N0 = 0;N0 < orb.Alpha[0].getNchi(L0);++N0)
+                                {
+                                    const int inl = this->inl_index[T0](I0, L0, N0);
+                                    const int nm = 2*L0+1;
+                                    for (int m1 = 0;m1 < 2 * L0 + 1;++m1)
+                                    {
+                                        for (int m2 = 0; m2 < 2 * L0 + 1; ++m2)
+                                        {
+                                            //(<d/dX chi_mu|alpha_m>)<chi_nu|alpha_m'>
+                                            gdmx[iat][inl][m1*nm+m2] += nlm2[1][ib+m2] * nlm1[ib+m1] * *dm_current;
+                                            gdmy[iat][inl][m1*nm+m2] += nlm2[2][ib+m2] * nlm1[ib+m1] * *dm_current;
+                                            gdmz[iat][inl][m1*nm+m2] += nlm2[3][ib+m2] * nlm1[ib+m1] * *dm_current;
+
+                                            //(<d/dX chi_nu|alpha_m'>)<chi_mu|alpha_m>
+                                            gdmx[iat][inl][m2*nm+m1] += nlm2[1][ib+m2] * nlm1[ib+m1] * *dm_current;
+                                            gdmy[iat][inl][m2*nm+m1] += nlm2[2][ib+m2] * nlm1[ib+m1] * *dm_current;
+                                            gdmz[iat][inl][m2*nm+m1] += nlm2[3][ib+m2] * nlm1[ib+m1] * *dm_current;                                            
+
+                                            //(<chi_mu|d/dX alpha_m>)<chi_nu|alpha_m'> = -(<d/dX chi_mu|alpha_m>)<chi_nu|alpha_m'>
+                                            gdmx[ibt2][inl][m1*nm+m2] -= nlm2[1][ib+m2] * nlm1[ib+m1] * *dm_current;         
+                                            gdmy[ibt2][inl][m1*nm+m2] -= nlm2[2][ib+m2] * nlm1[ib+m1] * *dm_current;         
+                                            gdmz[ibt2][inl][m1*nm+m2] -= nlm2[3][ib+m2] * nlm1[ib+m1] * *dm_current;
+
+                                            //(<chi_nu|d/dX alpha_m'>)<chi_mu|alpha_m> = -(<d/dX chi_nu|alpha_m'>)<chi_mu|alpha_m>
+                                            gdmx[ibt2][inl][m2*nm+m1] -= nlm2[1][ib+m2] * nlm1[ib+m1] * *dm_current; 
+                                            gdmy[ibt2][inl][m2*nm+m1] -= nlm2[2][ib+m2] * nlm1[ib+m1] * *dm_current;         
+                                            gdmz[ibt2][inl][m2*nm+m1] -= nlm2[3][ib+m2] * nlm1[ib+m1] * *dm_current; 
+
+                                            if (isstress)
+                                            {
+                                                int mm = 0;
+                                                for(int ipol=0;ipol<3;ipol++)
+                                                {
+                                                    for(int jpol=ipol;jpol<3;jpol++)
+                                                    {
+                                                        gdm_epsl[mm][inl][m2*nm+m1] += ucell.lat0 * 
+                                                        *dm_current * (nlm2[jpol+1][ib+m2] * nlm1[ib+m1] * r0[ipol]);
+                                                        mm++;
+                                                    }
+                                                }
+                                            }
+
+                                        }
+                                    }
+                                    ib+=nm;
+                                }
+                            }
+                            assert(ib==nlm1.size());
+
+                            if  (isstress)
+                            {
+                                nlm1 = this->nlm_save_k[iat][key_2][col_indexes[iw2l]][0];
+                                nlm2 = this->nlm_save_k[iat][key_1][row_indexes[iw1l]];
+
+                                assert(nlm1.size()==nlm2[0].size());  
+                                int ib=0;
+                                for (int L0 = 0; L0 <= orb.Alpha[0].getLmax();++L0)
+                                {
+                                    for (int N0 = 0;N0 < orb.Alpha[0].getNchi(L0);++N0)
+                                    {
+                                        const int inl = this->inl_index[T0](I0, L0, N0);
+                                        const int nm = 2*L0+1;
+                                        for (int m1 = 0;m1 < nm; ++m1)
+                                        {
+                                            for (int m2 = 0; m2 < nm; ++m2)
+                                            {
+                                                int mm = 0;
+                                                for(int ipol=0;ipol<3;ipol++)
+                                                {
+                                                    for(int jpol=ipol;jpol<3;jpol++)
+                                                    {
+                                                        gdm_epsl[mm][inl][m2*nm+m1]  += ucell.lat0 * 
+                                                        *dm_current * (nlm1[ib+m1] * nlm2[jpol+1][ib+m2] * r1[ipol]);
+                                                        mm++;
+                                                    }
+                                                }
+                                            }
+                                        }
+                                        ib+=nm;
+                                    }
+                                }
+                            }
+                            dm_current++;
+						}//iw2
+					}//iw1
+				}//ad2
+			}//ad1
+        }//I0
+    }//T0
+
+#ifdef __MPI
+    for(int iat=0;iat<ucell.nat;iat++)
+    {
+        allsum_deepks(this->inlmax,size,this->gdmx[iat]);
+        allsum_deepks(this->inlmax,size,this->gdmy[iat]);
+        allsum_deepks(this->inlmax,size,this->gdmz[iat]);
+    }
+    if (isstress)
+    {
+        for(int ipol=0;ipol<6;ipol++)
+        {
+            allsum_deepks(this->inlmax,size,this->gdm_epsl[ipol]);
+        }
+    }
+#endif
+    ModuleBase::timer::tick("LCAO_Deepks","cal_gdmx_k");
+    return;
+}
+
+#endif
